% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{trendy}
\alias{trendy}
\title{Google Trends Search}
\usage{
trendy(search_terms, from = NA, to = NA, ...)
}
\arguments{
\item{search_terms}{A character vector containing the search terms of interest}

\item{from}{The beginning date of the query}

\item{to}{The beginning date of the query}

\item{...}{arguments passed to \code{gtrendsR::gtrends()}. See ?gtrendsR::gtrends for more information including geography, language, and time-zone.}
}
\value{
An object of class `trendy`
}
\description{
Search Google Trends to retrieve relative hits and popularity. Note that this does not have the ability to compare trends as gtrendsR does (yet).
}
\examples{
\dontrun{
trendy("RStudio 1.2")
}

}
