% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R, R/zzz.R
\name{trendy}
\alias{trendy}
\alias{print.trendy}
\title{Google Trends Search}
\usage{
trendy(search_terms, from = NA, to = NA, ...)

\method{print}{trendy}(x, ...)
}
\arguments{
\item{search_terms}{A character vector containing the search terms of interest}

\item{from}{The beginning date of the query}

\item{to}{The beginning date of the query}

\item{...}{arguments passed to \code{gtrendsR::gtrends()}. See ?gtrendsR::gtrends for more information including geography, language, and time-zone.}

\item{x}{trendy object}
}
\value{
An object of class `trendy`
}
\description{
Search Google Trends to retrieve relative hits and popularity. This supports comparison of up to 5 search terms. Anything more than that will be searched individually.
}
\examples{
\dontrun{
trendy("RStudio 1.2")
}

}
